Part 8C: two searches on the Field, and what effort really costs
WHAT HE ASKED FOR

Two boxes above the chart. Type a provider and its models go green while the
rest of the field stays there in grey -- he said grey rather than hidden and he
was right, the greys are what make the green mean anything. Type a model and
only that family is drawn, its effort modes joined in order and each node
labelled with its MODE, because the model name is identical on all six and
would be six copies of one word. Both boxes carry a datalist; nobody types
`openai/gpt-5-6-terra-non-reasoning` from memory. Clearing either gives the
whole field back.

THE ANSWER TO HIS QUESTION IS "IT DEPENDS", AND THE CHART NOW SAYS ON WHAT

He wanted to see how effort affects price. On the posted price list it does not:
of the 10 families in these recordings that publish more than one mode, 6 charge
one rate for every one of them -- on his box 44 of 58. Those polylines are
vertical, and the screen says so in a sentence it counts from the data loaded
rather than one I typed in.

What moves is the cost of a task, because a harder mode returns more tokens at
the same rate. Only the gateway's own traffic knows how many, so the smoke store
now seeds some: gpt-6-astra runs 120 median output tokens at non-reasoning and
2400 at max, every mode at $10/$50 per million.

    posted price per 1M      six dots at $20.00, a vertical line, 5px wide
    cost per task            $0.026 at non-reasoning to $0.140 at max

A node measured from telemetry is a disc; one falling back to a posted price is
a hollow square, and the key sits beside the axis switch. The two are never
joined on one line without saying which is which.

THREE BUGS, AND ONE OF THEM WAS MINE FROM LAST NIGHT

A FOLD LEFT THE FAMILY POINTER BEHIND. `merge_pull` rewrites a model's id, its
aliases, its observations, its prices and its capabilities -- but not `family`,
and `family` is an id. So when AA's `openai/gpt-5-6-luna` folded onto
OpenRouter's `openai/gpt-5.6-luna`, every mode of that family went on pointing
at an id that no longer existed. 9 of 33 families; 4 of the 10 with more than
one mode. Nothing could find the family's base row, so the inheritance fix in
843bb74 reached none of them: tools arrived at 5 mode rows before, 23 after.
Your card told me to check that the nodes I draw are the ones the ranking can
seat. They were not. Two tests now pin it.

THE COST AXIS WAS A RACE. It merged all nine profiles' rankings and kept
whichever `cost_per_task` resolved first, so the same screen reloaded drew
different numbers. It now reads one named profile, alphabetical by default and
changeable, with the shape printed beside it. That turned out to matter more
than a tidy-up: the same six astra modes spread over 5.4x at cheap_bulk's 2,000
input tokens and converge to within 6% at reader's 200,000. "How much does
effort cost" has no answer until you say whose task.

THE X DOMAIN HAD A HARDCODED FLOOR. `Math.min(...xs, 0.0001)` folded a fallback
into the minimum, so the axis always reached down to $0.0001 whether or not
anything was there. A filtered field stayed squashed against the right edge and
a narrow domain drew a single tick label.

A CORRECTION TO MY OWN REPORT

I told you f777a5e was ruff clean. It was not -- one RUF046 in
sieve/scoring/leaderboard.py, a redundant int() around math.floor. CI on main
has been red since last night because of it. Fixed in this commit. I do not know
how I missed it; the honest answer is that I ran the linter before I wrote that
function and reported the earlier result.

212 pytest, 31 vitest, 41 playwright green. mypy --strict clean, ruff clean,
sieve check green, eslint / svelte-check / build clean.

Part 9 next, card t_bb756a1d: the fal tier parser first, then deepinfra.

+++ sieve/scoring/leaderboard.py
@@ -129,7 +129,7 @@ def _key(value: float) -> float:
     """
     if value == 0:
         return 0.0
-    places = max(1, 4 - int(math.floor(math.log10(abs(value)))) - 1)
+    places = max(1, 4 - math.floor(math.log10(abs(value))) - 1)
     return round(value, places)
